Rename betty.console.user to betty.rich.user
