+++ betty/rich/__init__.py
@@ -1,0 +1,3 @@
+"""
+Rich integration.
+"""
